Change type - double in project 002

+++ Project002(div)/Program.cs
@@ -1,9 +1,9 @@
 ﻿// Divide
 Console.Write("Введите делимое: ");
-int num1 = Convert.ToInt32(Console.ReadLine());
+double num1 = Convert.ToInt32(Console.ReadLine());
 Console.Write("Введите делитель: ");
-int num2 = Convert.ToInt32(Console.ReadLine());
-int res = num1 / num2;
+double num2 = Convert.ToInt32(Console.ReadLine());
+double res = num1 / num2;
 Console.Write("Результат деления: ");
 Console.WriteLine(res);
 Console.ReadKey();
